% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_generateFiltrationStats.R
\name{arbitrayFilterStats}
\alias{arbitrayFilterStats}
\title{Arbitray Filter Statistics}
\usage{
arbitrayFilterStats(object)
}
\arguments{
\item{object}{A circExperiment object}
}
\value{
A tibble with filtration statistics
}
\description{
Very simple helper function to import chimeric reads, and record which the quantity of included and excluded backsplice junction covering reads
}
\examples{
#stats <- arbitrayFilterStats(object = circObject)
}
