% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ensemblReleases.R
\name{availableEnsemblReleases}
\alias{availableEnsemblReleases}
\title{Find available Ensembl releases}
\usage{
availableEnsemblReleases(ahub, organism)
}
\arguments{
\item{organism}{Target species}

\item{ah}{An annotationHub object}
}
\value{
A character vector with the title of available EnsDb objects
}
\description{
This function subsets an AnnotationhHub object according to provided species,
and extracts the title of existing EnsDb objects.
}
\examples{
availableEnsemblReleases()

}
