% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_importMetadata.R
\name{checkMetadataFilepaths}
\alias{checkMetadataFilepaths}
\title{Check filepaths from metadata exists}
\usage{
checkMetadataFilepaths(meta)
}
\arguments{
\item{meta}{single row of a metadata data.frame (or tibble)}
}
\value{
A named logical vector containing exisitance
}
\description{
Check filepaths from metadata exists
}
