% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ensemblReleases.R
\name{determineExtremities}
\alias{determineExtremities}
\title{Determine avialable releases based on extremities}
\usage{
determineExtremities(ens_extremes, ref_extremes)
}
\arguments{
\item{ens_extremes}{A numeric vector with release numbers of the oldest and newest available annotation object}

\item{ref_extremes}{A numeric vector with release numbers of the oldest and newest available reference files}
}
\value{
A numeric vector with Ensembl releases, containing both annotation object and reference files
}
\description{
Determine avialable releases based on extremities
}
\examples{
#ah <- AnnotationHub()
#organism <- "Homo_sapiens"
#meta <- extractReferenceMeta(ah, organism)
#ref_extremes <- referenceExtremities(meta)
#ens_extremes <- EnsDbExtremities(queries)
#extremes <- determineExtremities(ens_extremes, ref_extremes)
#min_release <- extremes[1]
#max_release <- extremes[2]
}
