% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_referenceDownload.R
\name{downloadFile}
\alias{downloadFile}
\title{Download and decompress reference files}
\usage{
downloadFile(url, out_dir)
}
\arguments{
\item{url}{}

\item{out_dir}{}
}
\value{
Downloaded and unzipped file name
}
\description{
Download and decompress reference files
}
