% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ensemblReleases.R
\name{EnsDbExtremities}
\alias{EnsDbExtremities}
\title{Determine Ensembl release extremities
Extracts the minimal and maximum Ensembl release number}
\usage{
EnsDbExtremities(queries)
}
\arguments{
\item{queries}{A character vector consisting of titles for available EnsDb objects.}
}
\value{
A numeric vector of the length two, containing the oldest and latest available Ensembl annotation object
}
\description{
Determine Ensembl release extremities
Extracts the minimal and maximum Ensembl release number
}
\examples{
ah <- AnnotationHub()
organism <- "Homo_sapiens"
queries <- availableEnsemblReleases(ah, organism)
ens_extremes <- EnsDbExtremities(queries)
}
