% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ensemblReleases.R
\name{extractEnsemblReleaseNumerics}
\alias{extractEnsemblReleaseNumerics}
\title{Extract ensembl releases
Extracts the numeric value of Ensembl release numbers from the titles of EnsDb objects, as provided by `AnnotationHub`.}
\usage{
extractEnsemblReleaseNumerics(queries)
}
\arguments{
\item{queries}{A character vector consisting of titles for available EnsDb objects.}
}
\value{

}
\description{
Extract ensembl releases
Extracts the numeric value of Ensembl release numbers from the titles of EnsDb objects, as provided by `AnnotationHub`.
}
\examples{
ah <- AnnotationHub()
organism <- "Homo_sapiens"
queries <- availableEnsemblReleases(ah, organism)
releases <- extractEnsemblReleaseNumerics(queries)
}
