% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ensemblReleases.R
\name{extractReferenceMeta}
\alias{extractReferenceMeta}
\title{Extract reference data from Annotation object}
\usage{
extractReferenceMeta(ahub, organism, release = NULL)
}
\arguments{
\item{ahub}{}

\item{organism}{}
}
\value{

}
\description{
Extract reference data from Annotation object
}
\examples{
ah <- AnnotationHub()
organism <- "Homo_sapiens"
meta <- extractReferenceMeta(ah, organism)
}
