% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_IdxCommands.R
\name{generateSTARidx}
\alias{generateSTARidx}
\title{Generate STAR index}
\usage{
generateSTARidx(star, out_dir, fa_file, gtf_file, read_length, threads)
}
\arguments{
\item{threads}{}
}
\value{

}
\description{
Generate STAR index
}
\examples{

}
