% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_importMetadata.R
\name{importMetadata}
\alias{importMetadata}
\title{Import metadata
Import the uploaded metadata file and updates the colnames of the first three main columns.}
\usage{
importMetadata(meta_fn)
}
\arguments{
\item{meta_fn}{Full path to the uploaded metadata location}
}
\value{

}
\description{
Import metadata
Import the uploaded metadata file and updates the colnames of the first three main columns.
}
