% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_circTable.R
\name{makeTables}
\alias{makeTables}
\title{Make circRNA tables}
\usage{
makeTables(object, ah, circbase)
}
\arguments{
\item{object}{}

\item{ah}{}
}
\value{

}
\description{
This function generates a data table containing detected circRNAs
}
\examples{
ahdb <- AnnotationHub()[["AH79689"]] # Ensembl EnsDb Release 100
circs <- makeTable(object= circObject, ah = ahdb)

}
