% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_generateFiltrationStats.R
\name{plotFiltrationStats}
\alias{plotFiltrationStats}
\title{Plot filtration stats}
\usage{
plotFiltrationStats(object)
}
\arguments{
\item{object}{a circExperiment object}
}
\value{
Statistics on included and excluded reads after filtration
}
\description{
Plot filtration stats
}
\examples{
plotFiltrationStats(circObject)
}
