% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ensemblReleases.R
\name{referenceExtremities}
\alias{referenceExtremities}
\title{Determine avialable reference files release extermities}
\usage{
referenceExtremities(meta)
}
\arguments{
\item{meta}{}
}
\value{
A numeric vector of the length two, containing the oldest and latest available reference files
}
\description{
Determine avialable reference files release extermities
}
\examples{
#ah <- AnnotationHub()
#organism <- "Homo_sapiens"
#meta <- extractReferenceMeta(ah, organism)
#referenceExtremities(meta)
}
