% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_getAnnotation_supportedOrganism.R
\docType{data}
\name{supported_organisms}
\alias{supported_organisms}
\title{List of organisms available in Ensembl. 
Lists must be manually updated.}
\format{
An object of class \code{character} of length 2.
}
\usage{
supported_organisms
}
\description{
supported_organisms: Vector names are presented in the shiny app,
while values are used for internal function, and must thus be underscored
}
\details{
supported_builds: Vector names correspond to vector values of supported organisms,
while vector values correspond to genome builds
}
\keyword{datasets}
