% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_referenceDownload.R
\name{downloadFile}
\alias{downloadFile}
\title{Download and decompress reference files}
\usage{
downloadFile(url, out_dir)
}
\arguments{
\item{url}{Character with donwload URL}

\item{out_dir}{Character vector stating the full path to the output directory}
}
\value{
Downloaded and unzipped file name
}
\description{
Download and decompress reference files
}
\examples{
#ah <- AnnotationHub()
#organism <- "Homo_sapiens"
#build <- "GRCh38"
#release <- 100
#qry <- subset(ah, species == "Homo sapiens", rdataprovider == "Ensembl")
#meta <- mcols(qry)
#urls <- getDownloadLinks(meta, organism, build, release)
## Not run
#file1 <- downloadFile(url = urls[1], out_dir = "/tmp/file1"
}
