% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_referenceDownload.R
\name{getDownloadLinks}
\alias{getDownloadLinks}
\title{Find links for reference files}
\usage{
getDownloadLinks(meta, organism, build, release)
}
\arguments{
\item{organism}{}

\item{build}{}

\item{release}{}
}
\value{
Named list with GTF and FA URLs
}
\description{
Find links for reference files
}
